requirements.txt
fastapi[standard]>=0.115.0,<0.116.0
sqlalchemy[asyncio]>=2.0.0,<2.1.0
asyncpg>=0.29.0,<0.30.0
python-dotenv>=1.0.0,<1.1.0
alembic>=1.13.0,<1.14.0